requirements.txt
streamlit
python-dotenv
supabase
google-generativeai
PyMuPDF
PyPDF2
python-docx
regex
